+++ sdk/java/src/main/java/com/pulumi/kubernetes/ProviderArgs.java
@@ -136,16 +136,18 @@ public final class ProviderArgs extends com.pulumi.resources.ResourceArgs {
     }
 
     /**
-     * If present and set to false, disable Server-Side Apply mode.
+     * BETA FEATURE - If present and set to true, enable Server-Side Apply mode.
      * See https://github.com/pulumi/pulumi-kubernetes/issues/2011 for additional details.
+     * This feature is in developer preview, and is disabled by default.
      * 
      */
     @Import(name="enableServerSideApply", json=true)
     private @Nullable Output<Boolean> enableServerSideApply;
 
     /**
-     * @return If present and set to false, disable Server-Side Apply mode.
+     * @return BETA FEATURE - If present and set to true, enable Server-Side Apply mode.
      * See https://github.com/pulumi/pulumi-kubernetes/issues/2011 for additional details.
+     * This feature is in developer preview, and is disabled by default.
      * 
      */
     public Optional<Output<Boolean>> enableServerSideApply() {
@@ -471,8 +473,9 @@ public final class ProviderArgs extends com.pulumi.resources.ResourceArgs {
         }
 
         /**
-         * @param enableServerSideApply If present and set to false, disable Server-Side Apply mode.
+         * @param enableServerSideApply BETA FEATURE - If present and set to true, enable Server-Side Apply mode.
          * See https://github.com/pulumi/pulumi-kubernetes/issues/2011 for additional details.
+         * This feature is in developer preview, and is disabled by default.
          * 
          * @return builder
          * 
@@ -483,8 +486,9 @@ public final class ProviderArgs extends com.pulumi.resources.ResourceArgs {
         }
 
         /**
-         * @param enableServerSideApply If present and set to false, disable Server-Side Apply mode.
+         * @param enableServerSideApply BETA FEATURE - If present and set to true, enable Server-Side Apply mode.
          * See https://github.com/pulumi/pulumi-kubernetes/issues/2011 for additional details.
+         * This feature is in developer preview, and is disabled by default.
          * 
          * @return builder
          *