+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+- Add allowNullValues boolean option to pass Null values through helm configs without having them scrubbed (https://github.com/pulumi/pulumi-kubernetes/issues/2089)
+
+
 ## 3.22.0 (October 21, 2022)
 
 Important Note -- This release changes the Provider default to enable Server-Side Apply mode. This change is
@@ -10,7 +13,6 @@ for additional information about using Server-Side Apply with Pulumi's Kubernete
 
 - Fix values precedence in helm release (https://github.com/pulumi/pulumi-kubernetes/pull/2191)
 - Enable Server-Side Apply mode by default (https://github.com/pulumi/pulumi-kubernetes/pull/2206)
-- Add allowNullValues boolean option to pass Null values through helm configs without having them scrubbed (https://github.com/pulumi/pulumi-kubernetes/issues/2089)
 
 ## 3.21.4 (September 22, 2022)